# Dont sync carriage return lines