--- a/tests/test_sender.py
+++ b/tests/test_sender.py
@@ -29,7 +29,9 @@ from wandb.proto import wandb_internal_pb2 as pb
 
 
 def test_send_status_request(
-    mock_server, internal_sender, start_backend,
+    mock_server,
+    internal_sender,
+    start_backend,
 ):
     mock_server.ctx["stopped"] = True
     start_backend()
@@ -40,7 +42,9 @@ def test_send_status_request(
 
 
 def test_parallel_requests(
-    mock_server, internal_sender, start_backend,
+    mock_server,
+    internal_sender,
+    start_backend,
 ):
     mock_server.ctx["stopped"] = True
     work_queue = queue.Queue()
@@ -72,7 +76,11 @@ def test_parallel_requests(
 
 
 def test_resume_success(
-    mocked_run, test_settings, mock_server, internal_sender, start_backend,
+    mocked_run,
+    test_settings,
+    mock_server,
+    internal_sender,
+    start_backend,
 ):
     test_settings.resume = "allow"
     mock_server.ctx["resume"] = True
@@ -84,7 +92,11 @@ def test_resume_success(
 
 
 def test_resume_error_never(
-    mocked_run, test_settings, mock_server, internal_sender, start_backend,
+    mocked_run,
+    test_settings,
+    mock_server,
+    internal_sender,
+    start_backend,
 ):
     test_settings.resume = "never"
     mock_server.ctx["resume"] = True
@@ -98,7 +110,11 @@ def test_resume_error_never(
 
 
 def test_resume_error_must(
-    mocked_run, test_settings, mock_server, internal_sender, start_backend,
+    mocked_run,
+    test_settings,
+    mock_server,
+    internal_sender,
+    start_backend,
 ):
     test_settings.resume = "must"
     mock_server.ctx["resume"] = False
@@ -113,7 +129,11 @@ def test_resume_error_must(
 
 
 def test_save_live_existing_file(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     with open(os.path.join(mocked_run.dir, "test.txt"), "w") as f:
@@ -124,7 +144,11 @@ def test_save_live_existing_file(
 
 
 def test_save_live_write_after_policy(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "live")]})
@@ -135,7 +159,11 @@ def test_save_live_write_after_policy(
 
 
 def test_save_live_multi_write(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "live")]})
@@ -151,7 +179,11 @@ def test_save_live_multi_write(
 
 
 def test_save_live_glob_multi_write(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("checkpoints/*", "live")]})
@@ -185,7 +217,11 @@ def test_save_live_glob_multi_write(
 
 
 def test_save_rename_file(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "live")]})
@@ -201,7 +237,11 @@ def test_save_rename_file(
 
 
 def test_save_end_write_after_policy(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "end")]})
@@ -212,7 +252,11 @@ def test_save_end_write_after_policy(
 
 
 def test_save_end_existing_file(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     with open(os.path.join(mocked_run.dir, "test.txt"), "w") as f:
@@ -223,7 +267,11 @@ def test_save_end_existing_file(
 
 
 def test_save_end_multi_write(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "end")]})
@@ -239,7 +287,11 @@ def test_save_end_multi_write(
 
 
 def test_save_now_write_after_policy(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "now")]})
@@ -250,7 +302,11 @@ def test_save_now_write_after_policy(
 
 
 def test_save_now_existing_file(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     with open(os.path.join(mocked_run.dir, "test.txt"), "w") as f:
@@ -261,7 +317,11 @@ def test_save_now_existing_file(
 
 
 def test_save_now_multi_write(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "now")]})
@@ -277,7 +337,11 @@ def test_save_now_multi_write(
 
 
 def test_save_glob_multi_write(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("checkpoints/*", "now")]})
@@ -304,7 +368,11 @@ def test_save_glob_multi_write(
 
 
 def test_save_now_relative_path(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     internal_sender.publish_files({"files": [("foo/test.txt", "now")]})
@@ -318,7 +386,11 @@ def test_save_now_relative_path(
 
 
 def test_save_now_twice(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     file_path = os.path.join("foo", "test.txt")
@@ -335,8 +407,13 @@ def test_save_now_twice(
     print("DAMN DUDE", mock_server.ctx)
     assert len(mock_server.ctx["storage?file=foo/test.txt"]) == 2
 
+
 def test_output(
-    mocked_run, mock_server, internal_sender, start_backend, stop_backend,
+    mocked_run,
+    mock_server,
+    internal_sender,
+    start_backend,
+    stop_backend,
 ):
     start_backend()
     for i in range(100):
@@ -344,7 +421,10 @@ def test_output(
     internal_sender.publish_output("stdout", "\rFinal line baby\n")
     stop_backend()
     print("DUDE!", mock_server.ctx)
-    assert "Final line baby" in mock_server.ctx["file_stream"][0]["files"]["output.log"]["content"][0]
+    assert (
+        "Final line baby"
+        in mock_server.ctx["file_stream"][0]["files"]["output.log"]["content"][0]
+    )
 
 
 def test_upgrade_upgraded(