--- a/tests/test_sender.py
+++ b/tests/test_sender.py
@@ -28,11 +28,7 @@ from wandb.proto import wandb_internal_pb2
 from wandb.proto import wandb_internal_pb2 as pb
 
 
-def test_send_status_request(
-    mock_server,
-    internal_sender,
-    start_backend,
-):
+def test_send_status_request(mock_server, internal_sender, start_backend):
     mock_server.ctx["stopped"] = True
     start_backend()
 
@@ -41,11 +37,7 @@ def test_send_status_request(
     assert status_resp.run_should_stop
 
 
-def test_parallel_requests(
-    mock_server,
-    internal_sender,
-    start_backend,
-):
+def test_parallel_requests(mock_server, internal_sender, start_backend):
     mock_server.ctx["stopped"] = True
     work_queue = queue.Queue()
     start_backend()
@@ -76,11 +68,7 @@ def test_parallel_requests(
 
 
 def test_resume_success(
-    mocked_run,
-    test_settings,
-    mock_server,
-    internal_sender,
-    start_backend,
+    mocked_run, test_settings, mock_server, internal_sender, start_backend
 ):
     test_settings.resume = "allow"
     mock_server.ctx["resume"] = True
@@ -92,11 +80,7 @@ def test_resume_success(
 
 
 def test_resume_error_never(
-    mocked_run,
-    test_settings,
-    mock_server,
-    internal_sender,
-    start_backend,
+    mocked_run, test_settings, mock_server, internal_sender, start_backend
 ):
     test_settings.resume = "never"
     mock_server.ctx["resume"] = True
@@ -110,11 +94,7 @@ def test_resume_error_never(
 
 
 def test_resume_error_must(
-    mocked_run,
-    test_settings,
-    mock_server,
-    internal_sender,
-    start_backend,
+    mocked_run, test_settings, mock_server, internal_sender, start_backend
 ):
     test_settings.resume = "must"
     mock_server.ctx["resume"] = False
@@ -129,11 +109,7 @@ def test_resume_error_must(
 
 
 def test_save_live_existing_file(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     with open(os.path.join(mocked_run.dir, "test.txt"), "w") as f:
@@ -144,11 +120,7 @@ def test_save_live_existing_file(
 
 
 def test_save_live_write_after_policy(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "live")]})
@@ -159,11 +131,7 @@ def test_save_live_write_after_policy(
 
 
 def test_save_live_multi_write(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "live")]})
@@ -179,11 +147,7 @@ def test_save_live_multi_write(
 
 
 def test_save_live_glob_multi_write(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("checkpoints/*", "live")]})
@@ -217,11 +181,7 @@ def test_save_live_glob_multi_write(
 
 
 def test_save_rename_file(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "live")]})
@@ -237,11 +197,7 @@ def test_save_rename_file(
 
 
 def test_save_end_write_after_policy(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "end")]})
@@ -252,11 +208,7 @@ def test_save_end_write_after_policy(
 
 
 def test_save_end_existing_file(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     with open(os.path.join(mocked_run.dir, "test.txt"), "w") as f:
@@ -267,11 +219,7 @@ def test_save_end_existing_file(
 
 
 def test_save_end_multi_write(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "end")]})
@@ -287,11 +235,7 @@ def test_save_end_multi_write(
 
 
 def test_save_now_write_after_policy(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "now")]})
@@ -302,11 +246,7 @@ def test_save_now_write_after_policy(
 
 
 def test_save_now_existing_file(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     with open(os.path.join(mocked_run.dir, "test.txt"), "w") as f:
@@ -317,11 +257,7 @@ def test_save_now_existing_file(
 
 
 def test_save_now_multi_write(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("test.txt", "now")]})
@@ -337,11 +273,7 @@ def test_save_now_multi_write(
 
 
 def test_save_glob_multi_write(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("checkpoints/*", "now")]})
@@ -368,11 +300,7 @@ def test_save_glob_multi_write(
 
 
 def test_save_now_relative_path(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     internal_sender.publish_files({"files": [("foo/test.txt", "now")]})
@@ -386,11 +314,7 @@ def test_save_now_relative_path(
 
 
 def test_save_now_twice(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
+    mocked_run, mock_server, internal_sender, start_backend, stop_backend
 ):
     start_backend()
     file_path = os.path.join("foo", "test.txt")
@@ -408,13 +332,7 @@ def test_save_now_twice(
     assert len(mock_server.ctx["storage?file=foo/test.txt"]) == 2
 
 
-def test_output(
-    mocked_run,
-    mock_server,
-    internal_sender,
-    start_backend,
-    stop_backend,
-):
+def test_output(mocked_run, mock_server, internal_sender, start_backend, stop_backend):
     start_backend()
     for i in range(100):
         internal_sender.publish_output("stdout", "\rSome recurring line")