https://github.com/wandb/client/issues/1972

Description
-----------

We no longer sync carriage returns which makes syncing much faster.  Like LOTS faster:

**Before** (60m)

```shell
time wandb sync -e vanpelt -p slow-sync offline-run-20210316_144457-1zraaq85 
wandb: WARNING Found .wandb file, not streaming tensorboard metrics.
Syncing: https://wandb.ai/vanpelt/slow-sync/runs/1zraaq85 ...done.
wandb sync -e vanpelt -p slow-sync offline-run-20210316_144457-1zraaq85  60m 12.2s user 0.67s system 41% cpu 2401.158 total
```

**After** (8.45s)

```shell
vanpelt@cvp:~/Downloads|⇒  time wandb sync -e vanpelt -p slow-sync offline-run-20210316_144457-1zraaq85 
wandb: WARNING Found .wandb file, not streaming tensorboard metrics.
Syncing: https://wandb.ai/vanpelt/slow-sync/runs/1zraaq85 ...done.
wandb sync -e vanpelt -p slow-sync offline-run-20210316_144457-1zraaq85  8.45s user 0.67s system 81% cpu 11.158 total
```